app/requirements.txt
google-adk
google-cloud-aiplatform
aiohttp
asyncio
